Moved remaining classes to package `world`
- Moved remaining classes to package `world`
- Reduced visibility for `Robot`
